[WEAV-177] 유저 실 도메인 정의

+++ infrastructure/persistence/src/main/resources/db/migration/V8__create_user_sil.sql
@@ -2,7 +2,7 @@ create table user_sil
 (
     id      binary(16) not null,
     user_id binary(16) not null,
-    amount  int        not null,
+    amount  bigint     not null,
     primary key (id),
     index (user_id)
 )
